+++ cocos/2d/CCMenuItem.cpp
@@ -355,7 +355,7 @@ bool MenuItemFont::initWithString(const std::string& value, const ccMenuCallback
     _fontName = _globalFontName;
     _fontSize = _globalFontSize;
 
-    Label *label = Label::createWithSystemFont(value, _fontName, _fontSize);
+    Label *label = Label::createWithSystemFont(value, _fontName, (float)_fontSize);
     if (MenuItemLabel::initWithLabel(label, callback))
     {
         // do something ?
@@ -366,7 +366,7 @@ bool MenuItemFont::initWithString(const std::string& value, const ccMenuCallback
 void MenuItemFont::setFontSizeObj(int s)
 {
     _fontSize = s;
-    dynamic_cast<Label*>(_label)->setSystemFontSize(_fontSize);
+    dynamic_cast<Label*>(_label)->setSystemFontSize((float)_fontSize);
     this->setContentSize(dynamic_cast<Label*>(_label)->getContentSize());
 }
